naprawa dodawania i edyci jednostek

+++ app/src/main/java/pl/kalisz/ak/rafal/peczek/mojepomiary/jednostki/JednostkiEdytuj.java
@@ -2,16 +2,19 @@ package pl.kalisz.ak.rafal.peczek.mojepomiary.jednostki;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.Toast;
 
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.textfield.TextInputLayout;
 
 import java.util.Date;
@@ -28,8 +31,10 @@ public class JednostkiEdytuj extends AppCompatActivity {
     private TextInputLayout nazwa, wartosc;
     private AutoCompleteTextView dokladnosc, przeznaczenie;
     private TextInputLayout dokladnoscL, przeznaczenieL;
+    private Button aktualizuj, anuluj;
     private UsersRoomDatabase database;
     private Jednostka jednostka;
+    private int dokladnoscSelectedId, przeznaczenieSelectedId;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -41,6 +46,8 @@ public class JednostkiEdytuj extends AppCompatActivity {
 
         nazwa = (TextInputLayout) findViewById(R.id.editTextNazwaLayout);
         wartosc = (TextInputLayout) findViewById(R.id.editTextJednostkaLayout);
+        aktualizuj = (Button) findViewById(R.id.button_save_edit);
+        anuluj = (Button) findViewById(R.id.button_disable_edit);
         dokladnosc = (AutoCompleteTextView) findViewById(R.id.spinner);
         przeznaczenie = (AutoCompleteTextView) findViewById(R.id.spinner2);
         dokladnoscL = (TextInputLayout) findViewById(R.id.spinnerLayout);
@@ -51,22 +58,33 @@ public class JednostkiEdytuj extends AppCompatActivity {
         jednostka = database.localJednostkaDao().findById(jednostkaId);
         nazwa.getEditText().setText(jednostka.getNazwa());
         wartosc.getEditText().setText(jednostka.getWartosc());
+        dokladnoscSelectedId = jednostka.getDokladnosc();
+        przeznaczenieSelectedId = jednostka.getPrzeznaczenie();
         String[] listaDokladnosci = getResources().getStringArray(R.array.dokladnosc);
-        dokladnosc.setText(dokladnosc.getAdapter().getItem(jednostka.getDokladnosc()).toString(), true);
-        przeznaczenie.setText(przeznaczenie.getAdapter().getItem(jednostka.getPrzeznaczenie()).toString(), true);
+        dokladnosc.setText(dokladnosc.getAdapter().getItem(dokladnoscSelectedId).toString(), false);
+        przeznaczenie.setText(przeznaczenie.getAdapter().getItem(przeznaczenieSelectedId).toString(), false);
 
 
-        Button aktualizuj = (Button) findViewById(R.id.button_save_edit);
-        Button anuluj = (Button) findViewById(R.id.button_disable_edit);
+        dokladnosc.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                dokladnoscSelectedId = position;
+            }
+        });
+
+        przeznaczenie.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                przeznaczenieSelectedId = position;
+            }
+        });
+
         aktualizuj.setEnabled(false);
         anuluj.setEnabled(false);
         nazwa.setEnabled(false);
         wartosc.setEnabled(false);
-//        dokladnosc.setEnabled(false);
         dokladnoscL.setEnabled(false);
-//        przeznaczenie.setEnabled(false);
         przeznaczenieL.setEnabled(false);
-
 
     }
 
@@ -96,18 +114,30 @@ public class JednostkiEdytuj extends AppCompatActivity {
                     Toast.makeText(this, "Nie mozna edytować domyślnej jednostki", Toast.LENGTH_LONG).show();
                 }
                 return true;
-            case R.id.drop:
-                if(!jednostka.getCzyDomyslna()) {
-                    if (database != null) {
-                        Jednostka jednostka = database.localJednostkaDao().findById(jednostkaId);
-                        database.localJednostkaDao().delete(jednostka);
-                    }
-                    finish();
-                }
-                else {
+            case R.id.drop: {
+
+                if (!jednostka.getCzyDomyslna()) {
+                    MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(JednostkiEdytuj.this);
+                    builder.setMessage("Czy na pewno usunąć");
+//                builder.setTitle("Alert !");
+                    builder.setCancelable(false);
+                    builder.setPositiveButton("Tak", (DialogInterface.OnClickListener) (dialog, which) -> {
+                        if (database != null) {
+                            Jednostka jednostka = database.localJednostkaDao().findById(jednostkaId);
+                            database.localJednostkaDao().delete(jednostka);
+                        }
+                        finish();
+                    });
+
+                    builder.setNegativeButton("Nie", (DialogInterface.OnClickListener) (dialog, which) -> {
+                        dialog.cancel();
+                    });
+                    builder.show();
+                } else {
                     Toast.makeText(this, "Nie mozna usunąć domyślnej jednostki", Toast.LENGTH_LONG).show();
                 }
                 return true;
+            }
             default:
                 return super.onOptionsItemSelected(item);
         }
@@ -128,17 +158,15 @@ public class JednostkiEdytuj extends AppCompatActivity {
     public void aktualizujJednostke(View view) {
         String nazwa = this.nazwa.getEditText().getText().toString();
         String wartosc = this.wartosc.getEditText().getText().toString();
-        int dokladnosc = (int) this.dokladnosc.getText().length();
-        int przeznaczenie = (int) this.przeznaczenie.getText().length();
 
-        if(dokladnosc != 0 && przeznaczenie != 0 && nazwa.length() >= 2 && wartosc.length() >= 1) {
+        if(nazwa.length() >= 2 && wartosc.length() >= 1) {
             nazwa = nazwa.substring(0, 1).toUpperCase() + nazwa.substring(1).toLowerCase();
 
             Jednostka jednostka = database.localJednostkaDao().findById(jednostkaId);
             jednostka.setNazwa(nazwa);
             jednostka.setWartosc(wartosc);
-            jednostka.setDokladnosc(dokladnosc);
-            jednostka.setPrzeznaczenie(przeznaczenie);
+            jednostka.setDokladnosc(dokladnoscSelectedId);
+            jednostka.setPrzeznaczenie(przeznaczenieSelectedId);
             jednostka.setDataAktualizacji(new Date());
             database.localJednostkaDao().insert(jednostka);
             finish();
